intermediate WIP storage

+++ bobcat/irandstream/irandstream3.f
@@ -1,4 +1,4 @@
-inline IRandStream::IRandStream(int min, int max, long seed)
+inline IRandStream::IRandStream(int min, int max, size_t seed)
 :
     Randbuffer(min, max, seed),
     std::istream(this)
